Move to a new transpile mechanism
This allows an AST of the Greyscript language, and the first bits of turning a language (in this case, Python) into the AST, then turn the AST into Greyscript.

This also has a first cut of creating a method for transforming the AST by allowing deep inspection of the code.  This should allow for space-related optimizations, such as unifying shared code into functions, or adding debugging and logging notes for turning the errors in the transpiled version into locations in the source.